refactor: rename mod processor to handler
